refactor: modularize home feed screen and split components into files
